Moved sever jsx elements to their own component
